feat: add custom Electron binary support with SKIP_ELECTRON flag
- Modified dev-runner.js to skip auto-launching Electron when SKIP_ELECTRON env var is set
- Added proton:server script to start dev server without Electron
- Added proton script to launch custom Electron binary with debugging flags
- Prevents auto-restart of Electron on file changes when using custom binary

+++ .electron-vue/dev-runner.js
@@ -76,7 +76,7 @@ function startMain () {
 
       logStats('Main', stats)
 
-      if (electronProcess && electronProcess.kill) {
+      if (!process.env.SKIP_ELECTRON && electronProcess && electronProcess.kill) {
         manualRestart = true
         process.kill(electronProcess.pid)
         electronProcess = null
@@ -151,7 +151,12 @@ function init () {
 
   Promise.all([startRenderer(), startMain()])
     .then(() => {
-      startElectron()
+      if (!process.env.SKIP_ELECTRON) {
+        startElectron()
+      } else {
+        console.log(chalk.green('\n  Skipping Electron auto-start (SKIP_ELECTRON is set)'))
+        console.log(chalk.yellow('  You can now start Electron manually with your custom binary\n'))
+      }
     })
     .catch(err => {
       console.error(err)
